Condense GeoShape query tests into one file
This commit puts all GeoShape query test classes into one file since they are all related
to geo_shape queries. This also streamlines the generated documentation,
