--- docs/query-dsl/geo/shape/envelope/geo-shape-envelope-query-usage.asciidoc
@@ -1,45 +1,0 @@
-:ref_current: https://www.elastic.co/guide/en/elasticsearch/reference/6.1
-
-:github: https://github.com/elastic/elasticsearch-net
-
-:nuget: https://www.nuget.org/packages
-
-////
-IMPORTANT NOTE
-==============
-This file has been generated from https://github.com/elastic/elasticsearch-net/tree/master/src/Tests/QueryDsl/Geo/Shape/Envelope/GeoShapeEnvelopeQueryUsageTests.cs. 
-If you wish to submit a PR for any spelling mistakes, typos or grammatical errors for this file,
-please modify the original csharp file found at the link and submit the PR with that change. Thanks!
-////
-
-[[geo-shape-envelope-query-usage]]
-=== Geo Shape Envelope Query Usage
-
-==== Fluent DSL example
-
-[source,csharp]
-----
-q
-.GeoShapeEnvelope(c => c
-    .Name("named_query")
-    .Boost(1.1)
-    .Field(p=>p.Location)
-    .Coordinates(this._coordinates)
-    .Relation(GeoShapeRelation.Intersects)
-)
-----
-
-==== Object Initializer syntax example
-
-[source,csharp]
-----
-new GeoShapeEnvelopeQuery
-{
-    Name = "named_query",
-    Boost = 1.1,
-    Field = Field<Project>(p=>p.Location),
-    Shape = new EnvelopeGeoShape(this._coordinates),
-    Relation = GeoShapeRelation.Intersects,
-}
-----
-
