--- docs/query-dsl/geo/shape/geometry-collection/geo-shape-geometry-collection-query-usage.asciidoc
@@ -1,63 +1,0 @@
-:ref_current: https://www.elastic.co/guide/en/elasticsearch/reference/6.1
-
-:github: https://github.com/elastic/elasticsearch-net
-
-:nuget: https://www.nuget.org/packages
-
-////
-IMPORTANT NOTE
-==============
-This file has been generated from https://github.com/elastic/elasticsearch-net/tree/master/src/Tests/QueryDsl/Geo/Shape/GeometryCollection/GeoShapeGeometryCollectionQueryUsageTests.cs. 
-If you wish to submit a PR for any spelling mistakes, typos or grammatical errors for this file,
-please modify the original csharp file found at the link and submit the PR with that change. Thanks!
-////
-
-[[geo-shape-geometry-collection-query-usage]]
-=== Geo Shape Geometry Collection Query Usage
-
-==== Fluent DSL example
-
-[source,csharp]
-----
-q
-.GeoShapeGeometryCollection(c => c
-    .Name("named_query")
-    .Boost(1.1)
-    .Field(p=>p.Location)
-    .Geometries(
-        new PointGeoShape(this._pointCoordinates),
-        new MultiPointGeoShape(this._multiPointCoordinates),
-        new LineStringGeoShape(this._lineStringCoordinates),
-        new MultiLineStringGeoShape(this._multiLineStringCoordinates),
-        new PolygonGeoShape(this._polygonCoordinates),
-        new MultiPolygonGeoShape(this._multiPolygonCoordinates)
-    )
-    .Relation(GeoShapeRelation.Intersects)
-)
-----
-
-==== Object Initializer syntax example
-
-[source,csharp]
-----
-new GeoShapeGeometryCollectionQuery
-{
-    Name = "named_query",
-    Boost = 1.1,
-    Field = Infer.Field<Project>(p=>p.Location),
-    Shape = new Nest.GeometryCollection
-    {
-        Geometries = new IGeoShape[]
-        {
-            new PointGeoShape(this._pointCoordinates),
-            new MultiPointGeoShape(this._multiPointCoordinates),
-            new LineStringGeoShape(this._lineStringCoordinates),
-            new MultiLineStringGeoShape(_multiLineStringCoordinates),
-            new PolygonGeoShape(this._polygonCoordinates),
-            new MultiPolygonGeoShape(this._multiPolygonCoordinates),
-        }
-    },
-    Relation = GeoShapeRelation.Intersects,
-}
-----
-
