--- docs/query-dsl/geo/shape/indexed-shape/geo-shape-indexed-shape-query-usage.asciidoc
@@ -1,80 +1,0 @@
-:ref_current: https://www.elastic.co/guide/en/elasticsearch/reference/6.1
-
-:github: https://github.com/elastic/elasticsearch-net
-
-:nuget: https://www.nuget.org/packages
-
-////
-IMPORTANT NOTE
-==============
-This file has been generated from https://github.com/elastic/elasticsearch-net/tree/master/src/Tests/QueryDsl/Geo/Shape/IndexedShape/GeoShapeIndexedShapeQueryUsageTests.cs. 
-If you wish to submit a PR for any spelling mistakes, typos or grammatical errors for this file,
-please modify the original csharp file found at the link and submit the PR with that change. Thanks!
-////
-
-[[geo-shape-indexed-shape-query-usage]]
-=== Geo Shape Indexed Shape Query Usage
-
-The GeoShape IndexedShape Query supports using a shape which has already been indexed in another index and/or index type within a geoshape query.
-This is particularly useful for when you have a pre-defined list of shapes which are useful to your application and you want to reference this
-using a logical name (for example __New Zealand__), rather than having to provide their coordinates within the request each time.
-
-See the Elasticsearch documentation on {ref_current}/query-dsl-geo-shape-query.html[geoshape queries] for more detail.
-
-==== Fluent DSL example
-
-[source,csharp]
-----
-q
-.GeoIndexedShape(c => c
-    .Name("named_query")
-    .Boost(1.1)
-    .Field(p=>p.Location)
-    .IndexedShape(p=>p
-        .Id(2)
-        .Path(pp=>pp.Location)
-    )
-    .Relation(GeoShapeRelation.Intersects)
-)
-----
-
-==== Object Initializer syntax example
-
-[source,csharp]
-----
-new GeoIndexedShapeQuery
-{
-    Name = "named_query",
-    Boost = 1.1,
-    Field = Field<Project>(p=>p.Location),
-    IndexedShape = new FieldLookup
-    {
-        Id = 2,
-        Index = Index<Project>(),
-        Type = Type<Project>(),
-        Path = Field<Project>(p=>p.Location),
-    },
-    Relation = GeoShapeRelation.Intersects
-}
-----
-
-[source,javascript]
-.Example json output
-----
-{
-  "geo_shape": {
-    "_name": "named_query",
-    "boost": 1.1,
-    "location": {
-      "indexed_shape": {
-        "id": 2,
-        "type": "doc",
-        "index": "project",
-        "path": "location"
-      },
-      "relation": "intersects"
-    }
-  }
-}
-----
-
