+++ src/CodeGeneration/DocGenerator/StringExtensions.cs
@@ -161,11 +161,55 @@ namespace DocGenerator
 			{ "Script.Map", "\"if (doc['state'].value == \\\"Stable\\\") { params._agg.commits.add(doc['numberOfCommits'].value) }\"" },
 			{ "Script.Combine", "\"def sum = 0.0; for (c in params._agg.commits) { sum += c } return sum\"" },
 			{ "Script.Reduce", "\"def sum = 0.0; for (a in params._aggs) { sum += a } return sum\"" },
-			{ "this._polygonCoordinates", @"new[]{
+			{ "EnvelopeCoordinates", @"new [] { new [] { 45.0, -45.0 }, new [] { -45.0, 45.0 }}" },
+			{ "CircleCoordinates", @"new [] { 45.0, -45.0 }" },
+			{ "MultiPointCoordinates", @"new [] { new [] {38.897676, -77.03653}, new [] {38.889939, -77.009051} }" },
+			{ "MultiLineStringCoordinates", @"new[]
+											{
+												new [] { new [] { 2.0, 12.0 }, new [] { 2.0, 13.0 },new [] { 3.0, 13.0 }, new []{ 3.0, 12.0 } },
+												new [] { new [] { 0.0, 10.0 }, new [] { 0.0, 11.0 },new [] { 1.0, 11.0 }, new []{ 1.0, 10.0 } },
+												new [] { new [] { 0.2, 10.2 }, new [] { 0.2, 10.8 },new [] { 0.8, 10.8 }, new []{ 0.8, 12.0 } },
+											}" },
+			{ "MultiPolygonCoordinates", @"new[]
+											{
+												new []
+												{
+													new []
+													{
+														new [] { 10.0, -17.0},
+														new [] {15.0, 16.0},
+														new [] {0.0, 12.0},
+														new [] {-15.0, 16.0},
+														new [] { -10.0, -17.0},
+														new [] { 10.0, -17.0}
+													},
+													new []
+													{
+														new [] {8.2  , 18.2},
+														new [] { 8.2 , -18.8},
+														new [] { -8.8, -10.8},
+														new [] {8.8  , 18.2}
+													}
+												},
+												new []
+												{
+													new []
+													{
+														new [] { 8.0, -15.0},
+														new [] {15.0, 16.0},
+														new [] {0.0, 12.0},
+														new [] {-15.0, 16.0},
+														new [] { -10.0, -17.0},
+														new [] { 8.0, -15.0}
+													}
+												}
+											}" },
+			{ "PolygonCoordinates", @"new[]{
 										new []{ new [] {10.0, -17.0}, new [] {15.0, 16.0}, new [] {0.0, 12.0}, new [] {-15.0, 16.0}, new [] {-10.0, -17.0},new [] {10.0, -17.0}},
 										new []{ new [] {8.2, 18.2}, new [] {8.2, -18.8}, new [] {-8.8, -10.8}, new [] {8.8, 18.2}}
-									}"
-			}
+									}" },
+			{ "LineStringCoordinates", @"new [] { new [] {38.897676, -77.03653}, new [] {38.889939, -77.009051} }" },
+			{ "PointCoordinates", "new[] { 38.897676, -77.03653 }" },
 		};
 
 		public static bool TryGetJsonForAnonymousType(this string anonymousTypeString, out string json)
