--- src/Tests/QueryDsl/Geo/Shape/Point/GeoShapePointQueryUsageTests.cs
@@ -1,48 +1,0 @@
-﻿using Nest;
-using Tests.Framework.Integration;
-using Tests.Framework.ManagedElasticsearch.Clusters;
-using Tests.Framework.MockData;
-using static Nest.Infer;
-
-namespace Tests.QueryDsl.Geo.Shape.Point
-{
-	public class GeoShapePointQueryUsageTests : GeoShapeQueryUsageTestsBase
-	{
-		public GeoShapePointQueryUsageTests(ReadOnlyCluster i, EndpointUsage usage) : base(i, usage) { }
-
-		private readonly GeoCoordinate _coordinates = new[] { -77.03653, 38.897676 };
-
-		protected override object ShapeJson => new
-		{
-			type ="point",
-			coordinates = this._coordinates
-		};
-
-		protected override QueryContainer QueryInitializer => new GeoShapeQuery
-		{
-			Name = "named_query",
-			Boost = 1.1,
-			Field = Field<Project>(p=>p.Location),
-			Shape = new PointGeoShape(this._coordinates),
-			Relation = GeoShapeRelation.Intersects,
-		};
-
-		protected override QueryContainer QueryFluent(QueryContainerDescriptor<Project> q) => q
-			.GeoShape(c => c
-				.Name("named_query")
-				.Boost(1.1)
-				.Field(p=>p.Location)
-				.Shape(s => s
-					.Point(this._coordinates)
-				)
-				.Relation(GeoShapeRelation.Intersects)
-			);
-
-		protected override ConditionlessWhen ConditionlessWhen => new ConditionlessWhen<IGeoShapeQuery>(a => a.GeoShape)
-		{
-			q =>  q.Field = null,
-			q =>  q.Shape = null,
-			q =>  ((IPointGeoShape)q.Shape).Coordinates = null,
-		};
-	}
-}
